crear interfaz

+++ Examen/IAsignatura.cs
@@ -3,7 +3,7 @@
     public interface IAsignatura
     {
         double CalcularNotaFinal();
-        double CalcularNotaFinal(double n1, double n2, double n3);
+        double CalcularNotaFinal(int n1, int n2, int n3);
         string MensajeNotaFinal(double notaFinal);
         void Imprimir();
     }
